Add Intel's reproducibility_verifier.sh script
Since the script is used to not only verity Intel's prebuilt signed AEs,
but AEs built by a party that is being audited, some variables were
renamed, and some words changed to reflect this more general usage.

TODO: The git history was not preserved, and should be restored at some
point.